% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_operations.R
\name{gsub_fixed}
\alias{gsub_fixed}
\title{Wrapper for gsub to align arguments with stringi package}
\usage{
gsub_fixed(str, pattern, replacement)
}
\arguments{
\item{str}{character vector; strings to search in}

\item{pattern}{character vector with replacements for matched patterns}

\item{replacement}{character vector; search patterns}
}
\value{
a character vector
}
\description{
Wrapper for gsub to align arguments with stringi package
}
