% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_functions.R
\name{lfc}
\alias{lfc}
\title{List Files as a Named Character}
\usage{
lfc(.dirs, reg = "*", rec = FALSE)
}
\arguments{
\item{.dirs}{Paths to the directories}

\item{reg}{RegEx to find files (defaults to '*' all files)}

\item{rec}{Should the directories be searched recursively?}
}
\value{
A named character
}
\description{
List Files as a Named Character
}
