% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_functions.R
\name{pdf_convert}
\alias{pdf_convert}
\title{Wrapper around pdftools::pdf_convert}
\usage{
pdf_convert(
  .path_in,
  .dir_out,
  .format = "png",
  .pages = NULL,
  .dpi = 72,
  .antialias = TRUE,
  .opw = "",
  .upw = "",
  .verbose = TRUE,
  .pad = 4
)
}
\arguments{
\item{.path_in}{file path or raw vector with pdf data}

\item{.dir_out}{directory to store images}

\item{.format}{string with output format such as "png" or "jpeg". Must be equal to one of poppler_config()$supported_image_formats.}

\item{.pages}{vector with one-based page numbers to render. NULL means all pages.}

\item{.dpi}{resolution (dots per inch) to render}

\item{.antialias}{enable antialiasing. Must be "text" or "draw" or TRUE (both) or FALSE (neither).}

\item{.opw}{owner password}

\item{.upw}{user password}

\item{.verbose}{print some progress info to stdout}

\item{.pad}{how many leading zeros}
}
\value{
Images
}
\description{
Wrapper around pdftools::pdf_convert
}
