% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_functions.R
\name{pdf_decrypt}
\alias{pdf_decrypt}
\title{Decrypt PDFs}
\usage{
pdf_decrypt(.path_in, .path_out)
}
\arguments{
\item{.path_in}{full path to the pdf}

\item{.path_out}{full path to the new pdf file}
}
\value{
A pdf
}
\description{
Decrypt PDFs
}
