% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_functions.R
\name{pdf_images}
\alias{pdf_images}
\title{Extract PDF Images}
\usage{
pdf_images(
  .path_in,
  .dir_out,
  .convert = TRUE,
  .resize = Inf,
  .remove_dup = FALSE,
  .save = c(".rds", ".csv", ".fst"),
  .zip = FALSE,
  .time_outs = Inf,
  .max_imgs = Inf
)
}
\arguments{
\item{.path_in}{full path to the pdf}

\item{.dir_out}{directory}

\item{.convert}{convert images to .jpg (default = TRUE)}

\item{.resize}{resize large images (default 1000px for the longer dimension)}

\item{.remove_dup}{remove duplicated picture}

\item{.save}{output format of the info file}

\item{.zip}{zip images?}

\item{.time_outs}{Number of Seconds to wait for each process}

\item{.max_imgs}{Maximal Number of Images to process}
}
\value{
Images and Info Table
}
\description{
Extract PDF Images
}
