% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_functions.R
\name{pdf_images_convert}
\alias{pdf_images_convert}
\title{Convert Images to jpg}
\usage{
pdf_images_convert(.tab_info)
}
\arguments{
\item{.tab_info}{A Dataframe generated by pdf_images_extract()}
}
\value{
Converted Images and a Dataframe with Infos
}
\description{
Convert Images to jpg
}
