% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_functions.R
\name{pdf_images_duplicates}
\alias{pdf_images_duplicates}
\title{Deduplicate Images}
\usage{
pdf_images_duplicates(.tab_info)
}
\arguments{
\item{.tab_info}{A Dataframe generated by pdf_images_extract()}
}
\value{
Deleted Images and a Dataframe with Infos
}
\description{
Deduplicate Images
}
