% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_functions.R
\name{pdf_images_extract}
\alias{pdf_images_extract}
\title{Extract Images from PDF}
\usage{
pdf_images_extract(.path_in, .dir_out)
}
\arguments{
\item{.path_in}{Full Path to the PDF}

\item{.dir_out}{Full Path to the directory to save the images}
}
\value{
Images and Dataframe with Infos
}
\description{
Extract Images from PDF
}
