% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_functions.R
\name{pdf_images_log_table}
\alias{pdf_images_log_table}
\title{Write Logging Table}
\usage{
pdf_images_log_table(
  .tab_log,
  .start,
  .stop,
  .process,
  .error,
  .dir_main,
  .doc_id
)
}
\arguments{
\item{.tab_log}{Initial logging table}

\item{.start}{Start time}

\item{.stop}{Stop time}

\item{.process}{Process name}

\item{.error}{error name}

\item{.dir_main}{Main directory}

\item{.doc_id}{Document ID of the PDF}
}
\value{
A Dataframe
}
\description{
Write Logging Table
}
