% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_functions.R
\name{pdf_images_resize}
\alias{pdf_images_resize}
\title{Resize Images}
\usage{
pdf_images_resize(.tab_info, .resize = Inf)
}
\arguments{
\item{.tab_info}{A Dataframe generated by pdf_images_extract()}

\item{.resize}{Resized Images and a Dataframe with Infos}
}
\value{

}
\description{
Resize Images
}
