% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_functions.R
\name{pdf_images_save}
\alias{pdf_images_save}
\title{Save Image Info file}
\usage{
pdf_images_save(.tab_info, .doc_id, .save = c(".rds", ".csv", ".fst"))
}
\arguments{
\item{.tab_info}{A Dataframe generated by pdf_images_extract()}

\item{.doc_id}{Document ID of the PDF}

\item{.save}{c(".rds", ".csv", ".fst")}
}
\value{
Saved File
}
\description{
Save Image Info file
}
