% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_functions.R
\name{pdf_images_zip}
\alias{pdf_images_zip}
\title{Zip Image Folder}
\usage{
pdf_images_zip(.tab_info)
}
\arguments{
\item{.tab_info}{A Dataframe generated by pdf_images_extract()}
}
\value{
zipped Images
}
\description{
Zip Image Folder
}
