% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_functions.R
\name{pdf_info}
\alias{pdf_info}
\title{Wrapper around pdftools::pdf_info}
\usage{
pdf_info(.path_in)
}
\arguments{
\item{.path_in}{full path to the pdf}
}
\value{
A Dataframe
}
\description{
Wrapper around pdftools::pdf_info
}
