% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_functions.R
\name{pdf_rotate}
\alias{pdf_rotate}
\title{Rotate PDF Pages}
\usage{
pdf_rotate(.path_in, .path_out, .angle, .pages)
}
\arguments{
\item{.path_in}{full path to the pdf}

\item{.path_out}{full path to the new pdf file}

\item{.angle}{either a single value (-270, -180, -90, 90, 180, 270) or a integer vector of the same length as pages}

\item{.pages}{an integer vector}
}
\value{
a pdf
}
\description{
Rotate PDF Pages
}
