% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_functions.R
\name{pdf_to_txt}
\alias{pdf_to_txt}
\title{Convert PDF to txt with xpdftools}
\usage{
pdf_to_txt(.path_in = NULL, .path_out = NULL)
}
\arguments{
\item{.path_in}{full path to the pdf}

\item{.path_out}{full path to the new txt file}
}
\value{
A text file
}
\description{
Convert PDF to txt with xpdftools
}
