% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_functions.R
\name{pdf_to_txt_batch}
\alias{pdf_to_txt_batch}
\title{Batch Convert PDF to Text}
\usage{
pdf_to_txt_batch(
  .dir_in = NULL,
  .dir_out = NULL,
  .paths_in = NULL,
  .paths_out = NULL,
  .inst = 20
)
}
\arguments{
\item{.dir_in}{full path to the directory with the pdfs}

\item{.dir_out}{full path to the directory to store txt files}

\item{.paths_in}{if .dir_in = NULL, vector with file paths of the pdf}

\item{.paths_out}{if .dir_out = NULL, vector with file paths of the pdf}

\item{.inst}{max instances of pdftotex.exe to run}
}
\value{
text files
}
\description{
Batch Convert PDF to Text
}
