% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position_count.R
\name{position_count}
\alias{position_count}
\title{Get Position of Terms in Text}
\usage{
position_count(
  .table_terms,
  .table_text,
  .col_split_id = NULL,
  .workers = 1,
  .pre_check = 3
)
}
\arguments{
\item{.table_terms}{A Dataframe with at least 2 columns (tid: Term ID, term: Term) \cr
Term column must be a list column}

\item{.table_text}{A Dataframe with at least 2 columns (doc_id: A Document Identifier, text: Text) \cr
Text column mus be tokenized (one token per row)}

\item{.col_split_id}{Additionally an ID column that shows breaks in the Text Dataframe (e.g. an sentence ID)}

\item{.workers}{Number of workers (default = 1)}

\item{.pre_check}{Number of ngram pre-check}
}
\value{
A Dataframe
}
\description{
Get Position of Terms in Text
}
