% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_position_count_tables.R
\name{prepare_table_terms}
\alias{prepare_table_terms}
\title{Prepare Term List Table for position_count()}
\usage{
prepare_table_terms(.table_terms)
}
\arguments{
\item{.table_terms}{A Dataframe with at least 2 columns (tid: Term ID, term: Term) \cr}
}
\value{
A Dataframe with a term list column
}
\description{
Prepare Term List Table for position_count()
}
