% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_position_count_tables.R
\name{prepare_table_text}
\alias{prepare_table_text}
\title{Prepare Term List Table for position_count()}
\usage{
prepare_table_text(
  .table_text,
  .use_udpipe,
  .lan = "english-ewt",
  .dir_mod = getwd(),
  .tagger = c("default", "none"),
  .parser = c("none", "default"),
  .return_raw = FALSE
)
}
\arguments{
\item{.table_text}{A Dataframe with at least 2 columns (doc_id: A Document Identifier, text: Text) \cr}

\item{.use_udpipe}{Use UDPipe for tokenization (if FALSE the default, other parameters are not important)}

\item{.lan}{UDPipe language (see ?udpipe_download_model)}

\item{.dir_mod}{UDPipe model directory}

\item{.tagger}{one of c("default", "none")}

\item{.parser}{one of c("none", "default")}

\item{.return_raw}{return the UDPipe raw output}
}
\value{
A tokenized Dataframe
}
\description{
Prepare Term List Table for position_count()
}
