% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_in_string.R
\name{replace_words}
\alias{replace_words}
\title{Replace Words in Text/Tokens}
\usage{
replace_words(.string, .table_replace, .tokenized = FALSE, .regex = FALSE)
}
\arguments{
\item{.string}{character vector; strings to search in}

\item{.table_replace}{A Table with 2 columns (column names are not important) \cr
column 1: pattern word \cr
column 2: replace word \cr}

\item{.tokenized}{If input is already tokenized a more efficient algorithm is used. Gives a considerable speed advantage in large datasets}

\item{.regex}{If TRUE word boundaries are added, if FALSE blanks are added}
}
\value{
A character vector
}
\description{
Replace Words in Text/Tokens
}
\examples{
tab_replace <- tibble::tribble(
~ from, ~ to,
"A", "A1",
"B", "B1"
)

replace_words("A B", tab_replace)
replace_words("A B,", tab_replace, .regex = TRUE)
replace_words("A B", tab_replace, .tokenized = TRUE)
replace_words(c("A", "B"), tab_replace, .tokenized = TRUE)

}
