% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_operations.R
\name{string_operations}
\alias{string_operations}
\title{Perform multiple string operations}
\usage{
string_operations(.string, .table_string)
}
\arguments{
\item{.string}{character vector; strings to search in}

\item{.table_string}{Ordererd dataframe from RFtext package}
}
\value{
a character vector
}
\description{
Perform multiple string operations
}
