% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_operations.R
\docType{data}
\name{table_string}
\alias{table_string}
\title{Table containing arguments for string_operations()}
\format{
A Table with 5 columns:
\describe{
\item{name}{name of the operation}
\item{search}{search pattern}
\item{replace}{replace of the operation}
\item{example}{an example of the operation}
\item{call}{function call}
}
}
\usage{
table_string
}
\description{
Table containing arguments for string_operations()
}
\keyword{datasets}
