% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_functions.R
\name{xls_to_xlsx}
\alias{xls_to_xlsx}
\title{Convert XLS to XLSX}
\usage{
xls_to_xlsx(
  .dir_in = NULL,
  .dir_out = NULL,
  .paths_in = NULL,
  .paths_out = NULL,
  .inst = 20
)
}
\arguments{
\item{.dir_in}{full path to the directory with the xls}

\item{.dir_out}{full path to the directory to store xlsx files}

\item{.paths_in}{if .dir_in = NULL, vector with file paths of the xls}

\item{.paths_out}{if .dir_out = NULL, vector with file paths of the xlsx}

\item{.inst}{max instances of docto.exe to run}
}
\value{
xlsx file
}
\description{
Convert XLS to XLSX
}
