% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_functions.R
\name{doc_to_doc}
\alias{doc_to_doc}
\title{Wrapper Around docto.exe}
\usage{
doc_to_doc(.path_in, .path_out, .format)
}
\arguments{
\item{.path_in}{full path to the document}

\item{.path_out}{full path to the new document}

\item{.format}{format of conversion, see RFtext::table_doc_to_formats for available formats}
}
\value{
A document
}
\description{
Wrapper Around docto.exe
}
