% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_functions.R
\name{pdf_images}
\alias{pdf_images}
\title{Extract PDF Images}
\usage{
pdf_images(
  .path_in,
  .dir_out,
  .convert = TRUE,
  .resize = 1000,
  .remove_dup = FALSE,
  .save = c(".rds", ".csv", ".fst"),
  .zip = FALSE
)
}
\arguments{
\item{.path_in}{full path to the pdf}

\item{.dir_out}{directory}

\item{.convert}{convert images to .jpg (default = TRUE)}

\item{.resize}{resize large images (default 1000px for the longer dimension)}

\item{.remove_dup}{remove duplicated picture}

\item{.save}{output format of the info file}

\item{.zip}{zip images?}
}
\value{

}
\description{
Extract PDF Images
}
